chore: split action close tests in their own file
