+++ mergify_engine/tests/functional/actions/test_close.py
@@ -1,0 +1,39 @@
+# -*- encoding: utf-8 -*-
+#
+# Copyright © 2020 Mergify SAS
+#
+# Licensed under the Apache License, Version 2.0 (the "License"); you may
+# not use this file except in compliance with the License. You may obtain
+# a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+# WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+# License for the specific language governing permissions and limitations
+# under the License.
+import yaml
+
+from mergify_engine.tests.functional import base
+
+
+class TestCloseAction(base.FunctionalTestBase):
+    def test_close(self):
+        rules = {
+            "pull_request_rules": [
+                {
+                    "name": "rename label",
+                    "conditions": [f"base={self.master_branch_name}"],
+                    "actions": {"close": {"message": "WTF?"}},
+                }
+            ]
+        }
+
+        self.setup_repo(yaml.dump(rules))
+
+        p, _ = self.create_pr()
+
+        p.update()
+        self.assertEqual("closed", p.state)
+        self.assertEqual("WTF?", list(p.get_issue_comments())[-1].body)
